Do not re-download repositories

+++ downloader/downloader.sh
@@ -6,9 +6,15 @@ OLDIFS=$IFS
 IFS=','
 while read owner repo branch path folder
 do
-  curl -o /tmp/${repo}.zip https://github.com/${owner}/${repo}/archive/refs/heads/${branch}.zip -O -J -L
-  unzip /tmp/${repo}.zip -d /tmp
-  rm -f /tmp/${repo}.zip
+  DIR="/tmp/${repo}-${branch}"
+  if [ -d "$DIR" ]; then
+    echo "Using previous folder!"
+  else
+    curl -o /tmp/${repo}.zip https://github.com/${owner}/${repo}/archive/refs/heads/${branch}.zip -O -J -L
+    unzip /tmp/${repo}.zip -d /tmp
+    rm -f /tmp/${repo}.zip
+  fi
+  
   FOLDER=`echo $folder | tr ' ' '-' | tr '[:upper:]' '[:lower:]'`
   if [ "$FOLDER" = "public" ]; then
     rsync -a -v /tmp/${repo}-${branch}/${path} $SCRIPTPATH/../public
@@ -17,6 +23,10 @@ do
   else
     rsync -a -v /tmp/${repo}-${branch}/${path} $SCRIPTPATH/../pages/$FOLDER
   fi
+done <<< $INPUT
+
+while read owner repo branch
+do
   rm -rf /tmp/${repo}-${branch}
 done <<< $INPUT
 IFS=$OLDIFS
